--- app/Providers/EventServiceProvider.php
@@ -1,25 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Providers;
-
-use Illuminate\Foundation\Support\Providers\EventServiceProvider as ServiceProvider;
-
-class EventServiceProvider extends ServiceProvider
-{
-    /**
-     * The event listener mappings for the application.
-     *
-     * @var array<class-string,array<int,class-string>>
-     */
-    protected $listen = [];
-
-    /**
-     * Determine if events and listeners should be automatically discovered.
-     */
-    public function shouldDiscoverEvents(): bool
-    {
-        return true;
-    }
-}
